+++ src/net/sourceforge/kolmafia/swingui/table/IntegerRenderer.java
@@ -25,7 +25,7 @@ public class IntegerRenderer extends DefaultTableCellRenderer {
 
     int intValue =
         value instanceof Integer
-            ? ((Integer) value).intValue()
+            ? (Integer) value
             : StringUtilities.parseInt(value.toString());
 
     String stringValue = KoLConstants.COMMA_FORMAT.format(intValue);
